youcompleteme should take suggestions from vim-go
YouCompleteMe when compiled with gocode-completer takes the completion
from gocode directly. This uses the mdempsky/gocode which works fine
with projects which are in /home/uknth/Workspace/lib/Go, but doesn't
work for Go Modules. vim-go has gocode-gomod which works with go
modules, so the suggestions have to be passed on to vim-go instead of
YouCompleteMe using it's own engine.